+++ packages/antd-management-fast-framework/src/framework/DataMultiPageView/MultiPageDrawer/index.jsx
@@ -397,7 +397,7 @@ class MultiPageDrawer extends MultiPage {
           {this.showSearchForm && (searchForm || null) != null ? (
             <div style={{ flex: 0 }}>
               <Card
-                bordered={false}
+                variant="borderless"
                 className={styles.containorSearch}
                 styles={{
                   body: {
@@ -454,7 +454,7 @@ class MultiPageDrawer extends MultiPage {
                   // height: 'calc(100vh - 103px)',
                 },
               }}
-              bordered={false}
+              variant="borderless"
               className={styles.containorTable}
               extra={
                 <Space