+++ packages/antd-management-fast-framework/src/framework/DataSinglePageView/SinglePageDrawer/index.jsx
@@ -401,7 +401,7 @@ class SinglePageDrawer extends SinglePage {
           {this.showSearchForm && (searchForm || null) != null ? (
             <div style={{ flex: 0 }}>
               <Card
-                bordered={false}
+                variant="borderless"
                 className={styles.containorSearch}
                 styles={{
                   body: {
@@ -455,7 +455,7 @@ class SinglePageDrawer extends SinglePage {
                   overflow: 'hidden',
                 },
               }}
-              bordered={false}
+              variant="borderless"
               className={styles.containorTable}
               extra={
                 <Space