+++ packages/antd-management-fast-component/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 2.7.38
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-common@2.6.151
+
 ## 2.7.37
 
 ### Patch Changes